+++ pilz_industrial_motion_testutils/CHANGELOG.rst
@@ -6,6 +6,10 @@ Forthcoming
 -----------
 * update package dependencies
 * Add missing intialization
+* Add getter for CircJointInterimCart in XMLTestdataLoader
+
+0.4.1 (2019-02-27)
+------------------
 
 0.3.6 (2019-02-26)
 ------------------